Put spinner in the center of link dialogs
Without some CSS hacks, the spinner will appear in the top left of the
dialog due to the div being hidden when it's rendered, and therefore
sized at 0x0 pixels.

This change ensures it appears centered and adds some qualifying text to
let the user know what is being loaded. Currently only in German and
English; other language translations will need to be sourced still